Create README.md
Subject Introduction: The management and operation of hotels is a relatively common research object in the field of economics research. For the operation and management of the hotel, not only the quality of the hotel itself needs to be considered, but also the customer group cannot be ignored. This research project will focus on the factors that cause customers to cancel their reservations.
In this project, we use datasets from Kaggle..